[Feat] getProfiles from page -> Profile
