+++ src/pages/FromPage/FromPage.jsx
@@ -6,17 +6,14 @@ import theme from "../../styles/theme";
 import Profile from "../../components/Profile/Profile";
 import EditorContent from "../../components/Editor/Editor";
 import { showToast } from "../../components/common/Toast/Toast";
-import { getProfiles } from "../../api/profiles";
 import { postMessage } from "../../api/messages.api";
-import { useState, useEffect, useRef, useCallback } from "react";
+import { useState, useRef } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 
 const RELATIONSHIP = ["지인", "친구", "동료", "가족"];
 const FONT = ["Noto Sans", "Pretendard", "나눔명조", "나눔손글씨 손편지체"];
 
 export default function FromPage() {
-  const [isLoading, setIsLoading] = useState(false);
-  const [images, setImages] = useState([]);
   const [message, setMessage] = useState({
     sender: "",
     profileImageURL: null,
@@ -27,27 +24,6 @@ export default function FromPage() {
   const { id } = useParams();
   const navigate = useNavigate();
   const editorRef = useRef();
-
-  const getProfilesImg = useCallback(() => {
-    setIsLoading(true);
-    setTimeout(() => {
-      getProfiles()
-        .then((results) => {
-          setImages(results);
-          if (results.length > 0) {
-            setMessage((prev) => ({ ...prev, profileImageURL: results[0] }));
-          }
-        })
-        .catch((error) => console.error(error))
-        .finally(() => {
-          setIsLoading(false);
-        });
-    }, 2000);
-  }, []);
-
-  useEffect(() => {
-    getProfilesImg();
-  }, []);
 
   const handleChange = (name, value) => {
     setMessage((prev) => ({
@@ -73,8 +49,6 @@ export default function FromPage() {
       });
   };
 
-  console.log(isLoading);
-
   return (
     <S.FromContainer>
       <S.From>
@@ -89,9 +63,6 @@ export default function FromPage() {
         <S.Wrapper>
           <S.Label>프로필</S.Label>
           <Profile
-            images={images}
-            setImages={setImages}
-            isLoading={isLoading}
             onChange={(value) => handleChange("profileImageURL", value)}
           />
         </S.Wrapper>
